[REFACTOR] Use WP-CLI via docker image rather than composer dep.
Installing wp-cli via composer is fine, but it means we're at the mercy
of our other dependencies. i.e. we could hit a dependency version
conflict which could mean we couldn't install wp-cli or that other
library.

This uses the official docker image as part of our app image in much the
same way as we install composer itself.

Unfortunately (afaik) dependabot is unable to handle images within the
`COPY` command in docker, so we will need to keep the image up to date
ourselves.

Similarly, this commit also bumps the composer image we use to use the
latest version at time of writing (2.2.3).