[SETUP] Make use of docker-compose profiles.
Use [docker-compose profiles](https://docs.docker.com/compose/profiles/)
for starting up additional services such as blackfire for profiling and
backstop for visual-regression testing.

Unfortunately due us needing to extend nginx we can't use this for
styleguide due to the following not being support in v3 of
docker-compose:
https://docs.docker.com/compose/extends/#extending-services well we
probably still could but we'd need to bring up with the mutliple
docker-compose files as well so little point just yet.

https://docs.docker.com/compose/extends/#extending-services